requirements.txt
anyio==3.6.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
falcon==3.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
Nuitka==1.1.3
ordered-set==4.1.0
Pillow==9.2.0
python-dotenv==0.21.0
pytz==2022.4
PyYAML==6.0
requests==2.28.1
sniffio==1.3.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
websockets==10.3
